Bind query builder only in where statements
The query builder was using overloaded version of the knex client.
Each call to `client.queryBuilder()` was returning instance of the
extended builder.

This poses a problem - what will happen when the QB has global scopes
and it is used in wrapped `where` statements? Good example would be
`whereExists`. In most cases it references to a different table. With
globalScopes it could make a mess.

Now, the QueryBuilder will use own instance only for `whereWrapped`
statements. Knex uses it every time when one uses `where()` with a
callback.

Closes #11